app/infrastructure/src/authorizer/requirements.txt
PyJWT>=2.6.0
requests>=2.28.2
cryptography>=41.0.0